+++ src/prefect/core/flow.py
@@ -1131,8 +1131,13 @@ class Flow:
         if build:
             if not self.storage:
                 raise ValueError("This flow has no storage to build")
-            if self not in self.storage:
+            if self.name not in self.storage:
                 self.storage.add_flow(self)
+            else:
+                warnings.warn(
+                    "A flow with the same name is already contained in storage; if you changed your Flow since"
+                    " the last build, you might experience unexpected issues and should re-create your storage object."
+                )
             storage = self.storage.build()  # type: Optional[Storage]
         else:
             storage = self.storage