+++ tests/environments/storage/test_docker_storage.py
@@ -1,4 +1,5 @@
 import os
+import sys
 import tempfile
 from unittest.mock import MagicMock
 
@@ -35,16 +36,25 @@ def test_empty_docker_storage():
     storage = Docker()
 
     assert not storage.registry_url
-    assert not storage.base_image
+    assert storage.base_image.startswith("python:")
     assert not storage.image_name
     assert not storage.image_tag
     assert not storage.python_dependencies
     assert not storage.env_vars
     assert not storage.files
+    assert storage.prefect_version == "0.5.3"
     assert storage.base_url == "unix://var/run/docker.sock"
 
 
-def test_empty_docker_storage():
+@pytest.mark.parametrize("version_info", [(3, 5), (3, 6), (3, 7)])
+def test_docker_init_responds_to_python_version(monkeypatch, version_info):
+    version_mock = MagicMock(major=version_info[0], minor=version_info[1])
+    monkeypatch.setattr(sys, "version_info", version_mock)
+    storage = Docker()
+    assert storage.base_image == "python:{}.{}".format(*version_info)
+
+
+def test_initialized_docker_storage():
     storage = Docker(
         registry_url="test1",
         base_image="test3",
@@ -53,6 +63,7 @@ def test_empty_docker_storage():
         image_tag="test5",
         env_vars={"test": "1"},
         base_url="test_url",
+        prefect_version="master",
     )
 
     assert storage.registry_url == "test1"
@@ -62,6 +73,7 @@ def test_empty_docker_storage():
     assert storage.python_dependencies == ["test"]
     assert storage.env_vars == {"test": "1"}
     assert storage.base_url == "test_url"
+    assert storage.prefect_version == "master"
 
 
 def test_files_not_absolute_path():
@@ -183,6 +195,18 @@ def test_create_dockerfile_from_base_image():
             output = dockerfile.read()
 
         assert "FROM python:3.6" in output
+
+
+def test_create_dockerfile_from_prefect_version():
+    storage = Docker(prefect_version="master")
+
+    with tempfile.TemporaryDirectory() as tempdir:
+        storage.create_dockerfile_object(directory=tempdir)
+
+        with open(os.path.join(tempdir, "Dockerfile"), "r") as dockerfile:
+            output = dockerfile.read()
+
+        assert "prefect.git@master" in output
 
 
 def test_create_dockerfile_with_weird_flow_name():